moved _onCharacteristicSuccessRoll from actor-sheet to actor.  Fixed applying STR damage roll #3093
